Added tacle measurement interference and timer management

+++ src/hypercall.c
@@ -1,6 +1,7 @@
 #include <hypercall.h>
+#include <core.h>
 
-int hypercall(enum hypervisor_actions action, int arg0, int arg1, int arg2) {
+uint64_t hypercall(enum hypervisor_actions action, uint64_t arg0, uint64_t arg1, uint64_t arg2) {
     int hypercall_id = HYPERCALL_BASE_VALUE | action;
     int ret;
